move subdirectory into root; remove old files
