+++ .env
@@ -1,0 +1,6 @@
+API_KEY="AIzaSyCYin13y1r_WjpXGY7MK5wKq6BL1F8LEbw"
+AUTH_DOMAIN="mar-website-dev.firebaseapp.com"
+DATABASE_URL="https://mar-website-dev.firebaseio.com"
+PROJECT_ID="mar-website-dev"
+STORAGE_BUCKET="mar-website-dev.appspot.com"
+MESSAGING_SENDER_ID="128897541679"
