+++ .env.production
@@ -1,0 +1,5 @@
+API_KEY="AIzaSyDglCXmo4u9TcK85k_d_xO6Bh4xPWMMbEk"
+AUTH_DOMAIN="mar-website-30a21.firebaseapp.com"
+DATABASE_URL="https://mar-website-30a21.firebaseio.com"
+STORAGE_BUCKET="mar-website-30a21.appspot.com"
+MESSAGING_SENDER_ID="971613104370"
