--- Mar4/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "mar",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview --port 4173"
-  },
-  "dependencies": {
-    "lodash": "^4.17.21",
-    "moment": "^2.29.3",
-    "sass": "^1.52.3",
-    "vue": "^3.2.36",
-    "vue-router": "^4.0.15"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.3",
-    "vite": "^2.9.9"
-  }
-}
