--- Mar4/src/components/header.vue
@@ -1,15 +1,0 @@
-<template>
-
-	<header>
-
-		<nav>
-			<router-link to="/" exact>Home</router-link>
-			<router-link to="/shows" exact>Shows</router-link>
-			<router-link to="/music" exact>Music</router-link>
-			<router-link to="/merch" exact>Merch</router-link>
-			<router-link to="/info" exact>Info</router-link>
-		</nav>
-
-	</header>
-
-</template>
