--- Mar4/src/views/P404.vue
@@ -1,12 +1,0 @@
-<template>
-<transition appear>
-
-	<main>
-
-		<h1>Error 404</h1>
-		<p>Sorry but the page you requested was not found.</p>
-
-	</main>
-
-</transition>
-</template>
