+++ package.json
@@ -1,33 +1,20 @@
 {
-  "name": "mar4",
-  "description": "A Vue.js project",
-  "version": "1.0.0",
-  "author": "K Belardinelli <k@kangabell.co>",
-  "private": true,
+  "name": "mar",
+  "version": "0.0.0",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --hot",
-    "build": "cross-env NODE_ENV=production webpack --progress --hide-modules"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview --port 4173"
   },
   "dependencies": {
-    "dotenv-webpack": "^1.7.0",
-    "firebase": "^5.8.6",
-    "moment": "^2.29.2",
-    "vue": "^2.3.3",
-    "vue-router": "^3.0.2",
-    "vuefire": "^1.4.3"
+    "lodash": "^4.17.21",
+    "moment": "^2.29.3",
+    "sass": "^1.52.3",
+    "vue": "^3.2.36",
+    "vue-router": "^4.0.15"
   },
   "devDependencies": {
-    "babel-core": "^6.0.0",
-    "babel-loader": "^6.0.0",
-    "babel-preset-env": "^1.5.1",
-    "cross-env": "^3.0.0",
-    "css-loader": "^0.25.0",
-    "file-loader": "^0.9.0",
-    "node-sass": "^7.0.0",
-    "sass-loader": "^5.0.1",
-    "vue-loader": "^12.1.0",
-    "vue-template-compiler": "^2.3.3",
-    "webpack": "^2.6.1",
-    "webpack-dev-server": "^3.1.11"
+    "@vitejs/plugin-vue": "^2.3.3",
+    "vite": "^2.9.9"
   }
 }
