--- src/firebase.js
@@ -1,11 +1,0 @@
-import Firebase from 'firebase'
-
-const firebaseApp = Firebase.initializeApp({
-	apiKey: process.env.API_KEY,
-	authDomain: process.env.AUTH_DOMAIN,
-	databaseURL: process.env.DATABASE_URL,
-	storageBucket: process.env.STORAGE_BUCKET,
-	messagingSenderId: process.env.MESSAGING_SENDER_ID
-});
-
-export const db = firebaseApp.database();
