+++ src/views/P404.vue
