+++ style.scss
@@ -3,10 +3,17 @@
 
 $white: #fafbf7;
 $dark: #424c49;
-$black: #635C5C;
-$gray: lighten($black, 10);
+$black: #131416;
+$gray: #D0CBC6;
 $red: #a04c36;
 $tan: #D8C0A4;
+
+:root {
+	--color__background-body: #{$black};
+	--color__text-main: #{$gray};
+	--color__link: #{$gray};
+	--color__accent: #{$red};
+}
 
 $border-width: 0.125rem;
 
@@ -58,12 +65,12 @@ $spinSize: 40px;
 /*** LAYOUT ***/
 
 body {
-	background-color: $white;
+	background-color: var(--color__background-body);
 	margin: 0;
 	overflow-y: scroll; // keep scrollbar in place from page to page
 	position: relative;
 	&:before, &:after { // dark bars
-		background: $red;
+		background: var(--color__accent);
 		content: '';
 		height: 0.75rem;
 		position: fixed;
@@ -90,7 +97,7 @@ header {
 	width: 100%;
 	z-index: 1;
 	nav {
-		background-color: $white;
+		background-color: var(--color__background-body);
 		height: 4rem;
 		padding: 0.75rem 0;
 		text-align: center;
@@ -142,20 +149,20 @@ section.announcement .wrapper + .wrapper {
 
 @font-face {
   font-family: 'CG Omega';
-  src: url('/assets/fonts/cg-omega.woff2') format('woff2'),
-       url('/assets/fonts/cg-omega.woff') format('woff'),
-       url('/assets/fonts/cg-omega.ttf') format('truetype');
+  src: url('src/assets/fonts/cg-omega.woff2') format('woff2'),
+       url('src/assets/fonts/cg-omega.woff') format('woff'),
+       url('src/assets/fonts/cg-omega.ttf') format('truetype');
 }
 
 body {
-	color: $black;
+	color: var(--color__text-main);
 	font-family: Optima, 'CG Omega', 'Helvetica Neue', Helvetica, sans-serif;
 	font-size: 14px;
 	letter-spacing: 0.025em;
 	line-height: 1.75em;
 }
 
-h1, h2, h3, h4 {
+h1, h2, h3, h4, .emphasis {
 	font-weight: normal;
 	letter-spacing: 0.05em;
 	margin-top: 0;
@@ -175,9 +182,12 @@ p + h3 {
 }
 
 a {
-	color: $dark;
+	color: var(--color__link);
 	text-decoration: none;
 	transition: all 0.125s ease;
+}
+.link-img {
+	border: none;
 }
 .button {
 	border: $border-width solid ($tan);
@@ -187,7 +197,7 @@ a {
 	text-transform: uppercase;
 	&:hover,
 	&:focus {
-		border-color: $black;
+		border-color: var(--color__text-main);
 		cursor: pointer;
 	}
 }
@@ -201,18 +211,18 @@ nav a {
 	}
 	&:hover,
 	&:focus {
-		border-color: $dark;
+		border-color: var(--color__accent);
 	}
 }
 a.router-link-active {
-	border-color: $dark;
+	border-color: var(--color__accent);
 	cursor: default;
 }
 main a {
-	border-bottom: $border-width solid $tan;
+	border-bottom: $border-width solid $dark;
 	&:hover,
 	&:focus {
-		border-color: $dark;
+		border-color: var(--color__accent);
 	}
 }
 
@@ -253,7 +263,6 @@ small, p.small, figcaption {
 blockquote {
 	margin: 1rem 0;
 	max-width: 38rem;
-	min-height: 10rem;
 	position: relative;
 	@media only screen and (min-width: 600px) {
 		margin: 2rem 0;
@@ -261,13 +270,11 @@ blockquote {
 }
 
 ::selection {
-  background: lighten($tan, 10);
-  color: darken($black, 10);
+  background: $dark;
 }
 
 ::-moz-selection {
-  background: lighten($tan, 10);
-  color: darken($black, 10);
+  background: $dark;
 }
 
 iframe {
@@ -298,7 +305,7 @@ main.home {
 	    top: calc(50% - #{$spinSize/2});
 	    left: calc(50% - #{$spinSize/2});
 	}
-	img {
+	.logo {
 		margin: 2rem auto 0;
 		width: 200px;
 		@media only screen and (min-width: 600px) {
@@ -309,6 +316,15 @@ main.home {
 		}
 	}
 } // main.home
+
+.announcement {
+	@media only screen and (min-width: 600px) {
+		margin-top: 2rem;
+		img {
+			max-width: 46rem;
+		}
+	}
+}
 
 /*** SHOWS PAGE ***/
 
@@ -472,7 +488,7 @@ section.contact a.bandcamp {
 		border-bottom: $border-width solid $tan;
 	}
 	.icon:before {
-		background-image: url('../assets/img/bandcamp.png');
+		background-image: url('src/assets/img/bandcamp.png');
 		background-size: contain;
 		background-repeat: no-repeat;
 		content: '';
